requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
google-generativeai==0.8.5
google-genai==1.24.0
PyAudio==0.2.14
opencv-python==4.11.0.86
Pillow==11.3.0
numpy==2.3.1
websockets==15.0.1

